EY-4499 Tilgjengeliggjør brevmodeller
