+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/behandling/Personer.kt
@@ -3,7 +3,6 @@ package no.nav.etterlatte.brev.behandling
 import no.nav.etterlatte.brev.brevbaker.SoekerOgEventuellVerge
 import no.nav.etterlatte.libs.common.person.Verge
 import no.nav.pensjon.brevbaker.api.model.Foedselsnummer
-import java.time.LocalDate
 
 data class PersonerISak(
     val innsender: Innsender?,
@@ -26,10 +25,4 @@ data class Soeker(
     val under18: Boolean? = null,
     val foreldreloes: Boolean = false,
     val ufoere: Boolean = false,
-)
-
-data class Avdoed(
-    val fnr: Foedselsnummer,
-    val navn: String,
-    val doedsdato: LocalDate,
 )
