+++ libs/etterlatte-brev-model/build.gradle.kts
@@ -4,4 +4,5 @@ plugins {
 
 dependencies {
     implementation(project(":libs:saksbehandling-common"))
+    implementation("no.nav.pensjon.brevbaker:brevbaker-api-model-common:1.4.0")
 }
