+++ libs/etterlatte-brev-model/src/main/kotlin/no/nav/etterlatte/brev/model/BarnepensjonInformasjonDoedsfall.kt
@@ -1,20 +1,23 @@
-package no.nav.etterlatte.brev.model.bp
+package no.nav.etterlatte.brev.model
 
 import no.nav.etterlatte.brev.BrevDataRedigerbar
 import no.nav.etterlatte.brev.behandling.Avdoed
 
-data class BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt(
+data class BarnepensjonInformasjonDoedsfall(
     val avdoedNavn: String,
     val borIutland: Boolean,
+    val erOver18aar: Boolean,
 ) : BrevDataRedigerbar {
     companion object {
         fun fra(
             borIutland: Boolean,
+            erOver18aar: Boolean,
             avdoede: List<Avdoed>,
-        ): BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt =
-            BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt(
+        ): BarnepensjonInformasjonDoedsfall =
+            BarnepensjonInformasjonDoedsfall(
                 avdoedNavn = avdoede.first().navn,
                 borIutland = borIutland,
+                erOver18aar = erOver18aar,
             )
     }
 }
