+++ libs/etterlatte-brev-model/src/main/kotlin/no/nav/etterlatte/brev/model/BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt.kt
@@ -1,23 +1,20 @@
-package no.nav.etterlatte.brev.model.bp
+package no.nav.etterlatte.brev.model
 
 import no.nav.etterlatte.brev.BrevDataRedigerbar
 import no.nav.etterlatte.brev.behandling.Avdoed
 
-data class BarnepensjonInformasjonDoedsfall(
+data class BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt(
     val avdoedNavn: String,
     val borIutland: Boolean,
-    val erOver18aar: Boolean,
 ) : BrevDataRedigerbar {
     companion object {
         fun fra(
             borIutland: Boolean,
-            erOver18aar: Boolean,
             avdoede: List<Avdoed>,
-        ): BarnepensjonInformasjonDoedsfall =
-            BarnepensjonInformasjonDoedsfall(
+        ): BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt =
+            BarnepensjonInformasjonDoedsfallMellomAttenOgTjueVedReformtidspunkt(
                 avdoedNavn = avdoede.first().navn,
                 borIutland = borIutland,
-                erOver18aar = erOver18aar,
             )
     }
 }
