+++ libs/etterlatte-brev-model/src/main/kotlin/no/nav/etterlatte/brev/model/OmstillingsstoenadInformasjonDoedsfall.kt
@@ -1,4 +1,4 @@
-package no.nav.etterlatte.brev.model.oms
+package no.nav.etterlatte.brev.model
 
 import no.nav.etterlatte.brev.BrevDataRedigerbar
 import no.nav.etterlatte.brev.behandling.Avdoed
